requirements.txt
git+https://github.com/Rapptz/discord.py
python-dotenv~=1.0.0
SQLAlchemy